added README.md and made minor SEO improvements
I did not find the search for Jenkins and fasten came up high on the
list when using DuckDuckGo and Google so I added a specific landing
page which should be indexed higher as more pages link to it. More
content could go onto this page to help jump start Jenkins
installations with fastlane.